+++ shared/toadvillage.d.ts
@@ -1,35 +1,3 @@
-export interface ReqCard {
-  amount: number;
-  name: string;
-}
-
-export interface ScryfallPart {
-  component: string;
-  name: string;
-  uri: string;
-}
-
-export interface ScryfallCard {
-  all_parts?: ScryfallPart[];
-  color_identity: string[];
-  card_faces: {
-    image_uris: {
-      normal: string;
-    };
-    name: string;
-  }[];
-  image_uris: {
-    normal: string;
-  };
-  name: string;
-  type_line: string;
-}
-
-export interface MatchedCard {
-  amount: number;
-  card: ScryfallCard;
-}
-
 export interface Card {
   name: string;
   image: string;
