+++ utilities/toad-village-helper/types.d.ts
@@ -1,4 +1,4 @@
-import { FormattedCard, Card } from "../shared/toadvillage";
+import { FormattedCard } from "../../shared/toadvillage";
 
 export interface DownloadInput {
   commanders: FormattedCard[];
